fix(MarkdownTailwind): Fix the style in markdown blocks
Used a TailwindCSS plugin to control the style inside markdown block through the keyword 'prose'. This allows us to control the style of the different possible tags inside the markdown content

Requires an install with npm
