+++ icpc-frontned/package.json
@@ -15,6 +15,7 @@
     "@mdx-js/react": "^3.0.0",
     "@next/mdx": "^14.0.4",
     "@tailwindcss/forms": "^0.5.7",
+    "@tailwindcss/typography": "^0.5.15",
     "@types/chroma-js": "^2.4.3",
     "@types/katex": "^0.16.7",
     "@types/mdx": "^2.0.10",
