+++ icpc-frontned/src/app/exercises/[exercise]/page.tsx
@@ -1,5 +1,5 @@
 import ExerciseCardComponent from '@/app/components/cards/ExerciseCardComponent'
-import NewsBodyComponent from '@/app/components/panels/NewsBodyComponent'
+import MarkdownBodyComponent from '@/app/components/panels/MarkdownBodyComponent'
 import { TextComponent } from '@/app/components/text/TextComponent'
 import { enumTextTags } from '@/constants/types'
 import useExcerciseStore from '@/store/useExcerciseStore'
@@ -27,13 +27,13 @@ async function ExercisePage({ params }: Readonly<{ params: { exercise: string } 
       <ExerciseCardComponent exercise={exerciseBody} />
       <div className='mx-auto w-full max-w-7xl px-4 py-10 sm:px-6 lg:px-8 text-accent dark:text-dark-accent'>
         <TextComponent tag={enumTextTags.h1} >Descripción del problema:</TextComponent>
-        <NewsBodyComponent body={description.compiledSource} />
+        <MarkdownBodyComponent body={description.compiledSource} />
         <br />
         <TextComponent tag={enumTextTags.h1} >Restricciones:</TextComponent>
         <TextComponent> {exerciseBody.constraints} </TextComponent>
         <br />
         <TextComponent tag={enumTextTags.h1} >Solución del problema:</TextComponent>
-        <NewsBodyComponent body={solution.compiledSource} />
+        <MarkdownBodyComponent body={solution.compiledSource} />
       </div>
     </main>
   )
