+++ icpc-frontned/tailwind.config.ts
@@ -21,7 +21,8 @@ const config: Config = {
     }
   },
   plugins: [
-    require('@tailwindcss/forms')
+    require('@tailwindcss/forms'),
+    require('@tailwindcss/typography')
   ]
 
 }
